requirements.txt
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0
lxml>=4.9.0
